Choose Your Own Adventure Style Game
This is the beginning code for a choose your own adventure style game. It contains the class framework for future expansions and a partially completed story as a method in a class that will contain other hard-coded stories. In the future the program will contain file-IO reader and writer functions that a user can utilize to create their own story for friends to play.